chore: list files with git ls-files
